app/requirements.txt
Flask==2.3.3
Flask-WTF==1.2.1
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.5.4
PyJWT==2.8.0
WTForms==3.1.2
